src: message: remove checksum and add payload size
the messages will be sent over TCP which already guarantees that data
will arrive intact, so having a checksum in the protocol is not needed
and wasn't activated anyway.

as tcp data arrives as a stream, we need to know how many bytes to
receive when receiving. for this, we add a payload size to the
header, which the receiver can use by first receiving the header
bytes and then the payload bytes based on the header values

Signed-off-by: Mikkel Jakobsen <mikkel.aunsbjerg@prevas.dk>